Split profile views in separate files
