+++ profile/views/__init__.py
@@ -1,0 +1,3 @@
+from .activity import *
+from .user import *
+from .manage import *
